+++ versiondb/tsrocksdb/store.go
@@ -241,7 +241,9 @@ func (s Store) FixData(storeNames []string, dryRun bool) error {
 			return err
 		}
 	}
-
+	if dryRun {
+		return nil
+	}
 	return s.Flush()
 }
 
@@ -252,8 +254,11 @@ func (s Store) fixDataStore(storeName string, dryRun bool) error {
 		return err
 	}
 
-	batch := grocksdb.NewWriteBatch()
-	defer batch.Destroy()
+	var batch *grocksdb.WriteBatch
+	if !dryRun {
+		batch = grocksdb.NewWriteBatch()
+		defer batch.Destroy()
+	}
 
 	prefix := storePrefix(storeName)
 	readOpts := grocksdb.NewDefaultReadOptions()
@@ -262,14 +267,14 @@ func (s Store) fixDataStore(storeName string, dryRun bool) error {
 		realKey := cloneAppend(prefix, pair.Key)
 
 		readOpts.SetTimestamp(pair.Timestamp)
-		oldValue, err := s.db.GetCF(readOpts, s.cfHandle, realKey)
+		oldValue, oldTimestamp, err := s.db.GetCFWithTS(readOpts, s.cfHandle, realKey)
 		if err != nil {
 			return err
 		}
 
-		clean := bytes.Equal(oldValue.Data(), pair.Value)
+		clean := bytes.Equal(oldValue.Data(), pair.Value) && bytes.Equal(oldTimestamp.Data(), pair.Timestamp)
 		oldValue.Free()
-
+		oldTimestamp.Free()
 		if clean {
 			continue
 		}